+++ tests/app_runner/specs/test_app_spec.py
@@ -1,7 +1,7 @@
 import pytest
 import yaml
 
-from app_runner.specs.app.app_spec import AppVersion
+from app_runner.specs.app.app_version import AppVersion
 from app_runner.specs.app.commands_spec import CommandShell, CommandDocker, MountOptions, CommandsSpec
 from app_runner.specs.submitter_ref import SubmitterRef
